Get all posts from database to Home Page     1. Backend Part:         a. Created backend point named indexRouter in router folder and created a query to get all details of a specific post id         b. Add the router object in index.js (server)         c. Created new REST Client named 'client(index).rest' and then tested out the response         Response = all posts with all details of each post
    2. Frontend Part:
        a. Removed unncesary cards from index.html and made a copy on original index.html
        b. Then in the index.js catch the response from the backend and add those all posts into frontend as cards

Special note: Used a loop to generate cards in the frontend. The loop iterates over a list of post elements, creating a clickable card for each that, when clicked, navigates to post.html with the post's ID as a URL parameter, and dynamically manages the creation of new columns and grids for layout after every 2 and 4 cards, respectively.

Added lot of data to the sql as well